+++ src/runtime/tests/vm_tests.rs
@@ -950,7 +950,7 @@ f 1, 2, 3
         }
 
         #[test]
-        fn function_arg_unpacking_tuple() {
+        fn arg_unpacking_tuple() {
             let script = "
 f = |a, (_, c), d| a + c + d
 f 1, (2, 3), 4
@@ -959,7 +959,7 @@ f 1, (2, 3), 4
         }
 
         #[test]
-        fn function_arg_unpacking_tuple_nested() {
+        fn arg_unpacking_tuple_nested() {
             let script = "
 f = |a, (_, (c, d), _), f| a + c + d + f
 f 1, (2, (3, 4), 5), 6
@@ -968,7 +968,7 @@ f 1, (2, (3, 4), 5), 6
         }
 
         #[test]
-        fn function_arg_unpacking_list() {
+        fn arg_unpacking_list() {
             let script = "
 f = |a, [_, c], d| a + c + d
 f 1, [2, 3], 4
@@ -977,7 +977,7 @@ f 1, [2, 3], 4
         }
 
         #[test]
-        fn function_arg_unpacking_mixed() {
+        fn arg_unpacking_mixed() {
             let script = "
 f = |a, (b, [_, d]), e| a + b + d + e
 f 1, (2, [3, 4]), 5
@@ -986,13 +986,59 @@ f 1, (2, [3, 4]), 5
         }
 
         #[test]
-        fn function_arg_unpacking_with_capture() {
+        fn arg_unpacking_with_capture() {
             let script = "
 x = 10
 f = |a, (b, c)| a + b + c + x
 f 1, (2, 3)
 ";
             test_script(script, 16.into());
+        }
+
+        #[test]
+        fn arg_unpacking_ellipsis_at_end() {
+            let script = "
+f = |(a, b, ...)| a + b
+f (1, 2, 3, 4, 5)
+";
+            test_script(script, 3.into());
+        }
+
+        #[test]
+        fn arg_unpacking_ellipsis_with_id_at_end() {
+            let script = "
+f = |(a, b, others...)| a + b + others.size()
+f (1, 2, 3, 4, 5)
+";
+            test_script(script, 6.into());
+        }
+
+        #[test]
+        fn arg_unpacking_ellipsis_at_start() {
+            let script = "
+f = |(..., y, z)| y + z
+f (1, 2, 3, 4, 5)
+";
+            test_script(script, 9.into());
+        }
+
+        #[test]
+        fn arg_unpacking_ellipsis_with_id_at_start() {
+            let script = "
+f = |(others..., y, z)| y + z + others.size()
+f (1, 2, 3, 4, 5)
+";
+            test_script(script, 12.into());
+        }
+
+        #[test]
+        fn arg_unpacking_ellipsis_mixed() {
+            let script = "
+f = |[a, (tuple_others..., z), list_others...]|
+  a + list_others.sum() + tuple_others.size() + z
+f [10, (1, 2, 3), 20, 30]
+";
+            test_script(script, 65.into());
         }
 
         #[test]